requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
tensorflow-cpu==2.15.0
Pillow==10.1.0
python-multipart==0.0.6
numpy==1.24.3
tensorflowjs==4.13.0
